+++ packages/manager/apps/pci-ai-notebooks/src/hooks/api/ai/notebook/useStartNotebook.hook.tsx
@@ -20,7 +20,7 @@ export function useStartNotebook({
     },
     onError,
     onSuccess: () => {
-      // Invalidate service list query to get the latest data
+      // Invalidate notebooks list query to get the latest data
       queryClient.invalidateQueries({
         queryKey: [projectId, 'ai/notebook'],
       });
