+++ packages/manager/apps/pci-ai-notebooks/src/hooks/api/ai/notebook/useStopNotebook.hook.tsx
@@ -17,7 +17,7 @@ export function useStopNotebook({ onError, onStopSuccess }: UseStopNotebook) {
     },
     onError,
     onSuccess: () => {
-      // Invalidate service list query to get the latest data
+      // Invalidate notebooks list query to get the latest data
       queryClient.invalidateQueries({
         queryKey: [projectId, 'ai/notebook'],
       });
